+++ stanza/doc.go
@@ -20,18 +20,13 @@
 // users own types.
 // Stanzas created using either API are not guaranteed to be valid or enforce
 // specific stanza semantics.
-// For instance, using this package you could create an IQ without a unique ID,
-// which is illegal in XMPP.
-// Packages that require correct stanza semantics, such as the `mellium.im/xmpp`
-// package, are expected to enforce stanza semantics when encoding stanzas to a
-// stream.
 //
 // Custom Stanzas
 //
 // The stanza types in this package aren't very useful by themselves. To
 // transmit meaningful data our stanzas must contain a payload.
 // To add a payload with the struct based API we use composition to create a new
-// struct that contains the payload as additional fields.
+// struct where the payload is represented by additional fields.
 // For example, XEP-0199: XMPP Ping defines an IQ stanza with a payload named
 // "ping" qualified by the "urn:xmpp:ping" namespace.
 // To implement this in our own code we might create a Ping struct similar to